Add clock and fix a bug when quitting
If the Quit Game button was pressed before one second had passed
the game would not be able to clear the interval potentially allowing
multiple game clocks to run at the same time.  This fix allows the
interval to be accessed outside of the countDown function allowing all
intervals to be cleared upon quitting a game.